Moving to angular.js setup
