+++ src/dndapi/endpoints/donors.py
@@ -6,29 +6,17 @@ import json
 import dndapi.auth as auth
 import dndapi.database as database
 
-def to_json(d):
-    jso = {
-        'id': d['id'],
-        'first_name': d['first_name'],
-        'last_name': d['last_name'],
-        'email_address': d['email_address'],
-        'physical_address': d['physical_address']
-    }
-    if 'dci_number' in d:
-        jso['dci_number'] = d['dci_number']
-    return json.dumps(jso)
-
 def validate_donor_post(js):
     if ('id' not in js and 
-            'firstname' in js and
-            'lastname' in js and
-            'email' in js and
-            'address' in js):
+            'first_name' in js and
+            'last_name' in js and
+            'email_address' in js and
+            'physical_address' in js):
         return True
     else:
         return False
 
-@app.route('/api/donors/', methods=['POST',])
+@app.route('/api/donors', methods=['POST',])
 @app.route('/api/donors/<int:donor_id>', methods=['GET'])
 @jwt_required()
 def get_donors(donor_id=None):
@@ -39,7 +27,7 @@ def get_donors(donor_id=None):
             d = database.get_donor_by_id(donor_id)
             app.logger.info(d)
             if d:
-                return to_json(d), 200, {'Content-Type': 'application/json; charset=utf-8'}
+                return json.dumps(d), 200, {'Content-Type': 'application/json; charset=utf-8'}
             else:
                 return '{"error": "donor not found"}', 404, {'Content-Type': 'application/json; charset=utf-8'}
         else:
@@ -48,18 +36,20 @@ def get_donors(donor_id=None):
         # pull the posted information from json and validate it
         #app.logger.info("method was post")
         json_data = request.get_json()
+        app.logger.info(json_data)
         ##app.logger.debug("json_data = %s", json_data)
         if not json_data or not validate_donor_post(json_data):
+            app.logger.info('/api/donors POST failed validation')
             return '', 400
         else:
             # make dci availabe if not entered
-            if 'dci' not in json_data:
-                json_data['dci'] = None
+            if 'dci_number' not in json_data:
+                json_data['dci_number'] = None
             # insert the new donor
             app.logger.info(json_data)
-            ins_donor = database.insert_new_donor(json_data['firstname'],
-                                      json_data['lastname'],
-                                      json_data['address'],
-                                      json_data['dci'],
-                                      json_data['email'] )
-            return to_json(ins_donor), 201, {'Content-Type': 'application/json; charset=utf-8'}
+            ins_donor = database.insert_new_donor(json_data['first_name'],
+                                      json_data['last_name'],
+                                      json_data['physical_address'],
+                                      json_data['dci_number'],
+                                      json_data['email_address'] )
+            return json.dumps(ins_donor), 201, {'Content-Type': 'application/json; charset=utf-8'}
