+++ src/dndapi/static/css/dnd.css
@@ -1,0 +1,5 @@
+.container {
+    border: 1px solid black;
+    padding: 10px;
+    margin: 10px;
+}
