--- src/dndapi/static/newindex.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" ng-app="dnd">
-  <head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.7.8/angular.min.js"></script>
-    <script src="js/newdndonations.js"></script>
-  </head>
-  <body>
-    <div ng-controller="LoginController">
-      <form novalidate>
-          <label>Username: <input type="text" ng-model="user.username"/></label><br/>
-          <label>Password: <input type="password" ng-model="user.password"/></label><br/>
-          <input type="submit" ng-click="doLogin(user)" value="Login" />
-      </form>
-      <div>{{ user | json }}</div>
-      <div>{{ apikey }}</div>
-    </div>
-  </body>
-</html>
